ATT-41:Fixing formating changes

+++ omod-1.10/src/main/java/org/openmrs/module/attachments/rest/AttachmentResource1_10.java
@@ -102,7 +102,7 @@ public class AttachmentResource1_10 extends DataDelegatingCrudResource<Attachmen
 		String instructions = context.getParameter("instructions");
 		
 		// Verify File Size
-		if (attachmentsContext.getMaxUploadFileSize() * 1024 * 1024 < file.getSize()) {
+		if (attachmentsContext.getMaxUploadFileSize() * 1024 * 1024 < (double)file.getSize()) {
 			throw new IllegalRequestException("The file  exceeds the maximum size");
 		}
